+++ programs/hodlhub/src/processors/initialize.rs
@@ -5,7 +5,7 @@ pub fn exec(
   ctx: Context<Initialize>,
   operators: Vec<Pubkey>,
   sol_target: u64,
-  protocol_fee_bps: u16,
+  protocol_fee_bps: u64,
 ) -> Result<()> {
   let owner = ctx.accounts.owner.key();
   let state = &mut ctx.accounts.state;
